feat: consolidate placeholder audit script
