+++ scripts/dashboard_placeholder_sync.py
@@ -22,7 +22,7 @@ def sync(dashboard_dir: Path, analytics_db: Path) -> None:
     count = 0
     if analytics_db.exists():
         with sqlite3.connect(analytics_db) as conn:
-            cur = conn.execute("SELECT COUNT(*) FROM placeholder_audit")
+            cur = conn.execute("SELECT COUNT(*) FROM todo_fixme_tracking")
             count = cur.fetchone()[0]
 
     compliance = max(0, 100 - count)
