--- scripts/placeholder_audit_logger.py
@@ -1,225 +1,0 @@
-"""Placeholder Audit Logger.
-
-This script scans the workspace for unfinished code markers such as
-``TODO`` and ``FIXME``. All findings are recorded in ``analytics.db``
-under the ``placeholder_audit`` table and a summary is written to the
-``dashboard/compliance`` directory.
-
-The implementation follows the database-first pattern by fetching
-additional placeholder patterns from ``production.db``. Visual progress
-is displayed with ``tqdm`` progress bars. The script is designed for
-non-interactive use and adheres to the project's enterprise guidelines.
-"""
-
-from __future__ import annotations
-
-import json
-import logging
-import os
-import re
-import sqlite3
-import time
-from datetime import datetime
-from pathlib import Path
-from typing import Iterable, List
-
-from tqdm import tqdm
-
-from scripts.database.add_code_audit_log import ensure_code_audit_log
-
-TEXT = {
-    "start": "[START]",
-    "success": "[SUCCESS]",
-    "error": "[ERROR]",
-    "info": "[INFO]",
-    "progress": "[PROGRESS]",
-}
-
-
-def fetch_db_patterns(production_db: Path) -> List[str]:
-    """Fetch placeholder patterns stored in ``production.db``."""
-    if not production_db.exists():
-        return []
-    with sqlite3.connect(production_db) as conn:
-        try:
-            cur = conn.execute("SELECT placeholder_name FROM template_placeholders")
-            return [row[0] for row in cur.fetchall()]
-        except sqlite3.Error as exc:  # pragma: no cover - log only
-            logging.error(f"{TEXT['error']} failed to read production.db: {exc}")
-            return []
-
-
-def _severity(pattern: str) -> str:
-    pat = pattern.lower()
-    if pat in {"todo", "pass"}:
-        return "low"
-    if pat == "fixme":
-        return "medium"
-    return "high"
-
-
-def scan_files(workspace: Path, patterns: Iterable[str]) -> List[dict]:
-    """Return a list of placeholder findings in ``workspace``."""
-    results: List[dict] = []
-    files = [p for p in workspace.rglob("*") if p.is_file()]
-    with tqdm(total=len(files), desc=f"{TEXT['progress']} scanning", unit="file") as bar:
-        for file in files:
-            try:
-                lines = file.read_text(encoding="utf-8", errors="ignore").splitlines()
-            except Exception:  # pragma: no cover - log only
-                bar.update(1)
-                continue
-            for line_num, line in enumerate(lines, 1):
-                for pat in patterns:
-                    if re.search(pat, line):
-                        results.append(
-                            {
-                                "file": str(file),
-                                "pattern": pat,
-                                "line": line_num,
-                                "context": line.strip()[:200],
-                            }
-                        )
-            bar.update(1)
-    return results
-
-
-def log_results(results: List[dict], db_path: Path) -> None:
-    """Insert placeholder findings into ``analytics.db`` with progress bars."""
-    db_path.parent.mkdir(parents=True, exist_ok=True)
-    ensure_code_audit_log(db_path)
-    with sqlite3.connect(db_path) as conn, tqdm(
-        total=len(results), desc=f"{TEXT['progress']} logging", unit="item"
-    ) as bar:
-        conn.execute(
-            """
-            CREATE TABLE IF NOT EXISTS placeholder_audit (
-                id INTEGER PRIMARY KEY,
-                file_path TEXT,
-                pattern TEXT,
-                line INTEGER,
-                severity TEXT,
-                ts TEXT
-            )
-            """,
-        )
-        for row in results:
-            conn.execute(
-                "INSERT INTO placeholder_audit (file_path, pattern, line, severity, ts) VALUES (?, ?, ?, ?, ?)",
-                (
-                    row["file"],
-                    row["pattern"],
-                    row["line"],
-                    _severity(row["pattern"]),
-                    datetime.now().isoformat(),
-                ),
-            )
-            conn.execute(
-                "INSERT INTO code_audit_log (file_path, line_number, placeholder_type, context, timestamp)"
-                " VALUES (?, ?, ?, ?, ?)",
-                (
-                    row["file"],
-                    row["line"],
-                    row["pattern"],
-                    row["context"],
-                    datetime.now().isoformat(),
-                ),
-            )
-            bar.update(1)
-        conn.commit()
-
-
-def rollback_last_entry(db_path: Path) -> bool:
-    """Remove the most recent placeholder audit entry from both tables."""
-    if not db_path.exists():
-        return False
-    removed = False
-    with tqdm(total=1, desc=f"{TEXT['progress']} rollback", unit="entry") as bar:
-        with sqlite3.connect(db_path) as conn:
-            cur = conn.execute("SELECT rowid FROM placeholder_audit ORDER BY rowid DESC LIMIT 1")
-            row = cur.fetchone()
-            if row:
-                conn.execute("DELETE FROM placeholder_audit WHERE rowid = ?", (row[0],))
-                conn.execute(
-                    "DELETE FROM code_audit_log WHERE rowid = ("
-                    "SELECT rowid FROM code_audit_log ORDER BY rowid DESC LIMIT 1)"
-                )
-                conn.commit()
-                removed = True
-            bar.update(1)
-    return removed
-
-
-def update_dashboard(results: List[dict], dashboard_dir: Path) -> None:
-    """Write a summary JSON file with compliance metrics."""
-    dashboard_dir.mkdir(parents=True, exist_ok=True)
-    count = len(results)
-    compliance = max(0, 100 - count)
-    status = "complete" if count == 0 else "issues_pending"
-    data = {
-        "timestamp": datetime.now().isoformat(),
-        "findings": count,
-        "compliance_score": compliance,
-        "progress_status": status,
-    }
-    (dashboard_dir / "placeholder_summary.json").write_text(json.dumps(data, indent=2), encoding="utf-8")
-
-
-def main(
-    workspace_path: str | None = None,
-    analytics_db: str | None = None,
-    production_db: str | None = None,
-    dashboard_dir: str | None = None,
-    simulate: bool = False,
-) -> bool:
-    """Run the placeholder audit logger.
-
-    Parameters
-    ----------
-    simulate:
-        If ``True``, skip writing to the database and dashboard.
-    """
-    logging.basicConfig(level=logging.INFO, format="%(message)s")
-    logging.info(f"{TEXT['start']} placeholder audit")
-
-    workspace = Path(workspace_path or os.getenv("GH_COPILOT_WORKSPACE", Path.cwd()))
-    analytics = Path(analytics_db or workspace / "databases" / "analytics.db")
-    production = Path(production_db or workspace / "databases" / "production.db")
-    dashboard = Path(dashboard_dir or workspace / "dashboard" / "compliance")
-
-    patterns = ["TODO", "FIXME", "Implementation placeholder"] + fetch_db_patterns(production)
-
-    start = time.time()
-    results = scan_files(workspace, patterns)
-    # Support test-mode via environment variable for automated runs
-    simulate = simulate or os.getenv("GH_COPILOT_TEST_MODE") == "1"
-
-    if not simulate:
-        log_results(results, analytics)
-        update_dashboard(results, dashboard)
-    else:
-        logging.info("[TEST MODE] Simulation enabled: no database writes")
-    elapsed = time.time() - start
-    logging.info(f"{TEXT['success']} audit completed in {elapsed:.2f}s")
-
-    return True
-
-
-if __name__ == "__main__":  # pragma: no cover - CLI execution
-    import argparse
-
-    parser = argparse.ArgumentParser(description="Log placeholder audit results")
-    parser.add_argument("--workspace-path", type=str, help="Workspace to scan")
-    parser.add_argument("--analytics-db", type=str, help="Path to analytics.db")
-    parser.add_argument("--production-db", type=str, help="Path to production.db")
-    parser.add_argument("--dashboard-dir", type=str, help="Compliance dashboard directory")
-    parser.add_argument("--simulate", action="store_true", help="Test mode")
-    args = parser.parse_args()
-    success = main(
-        workspace_path=args.workspace_path,
-        analytics_db=args.analytics_db,
-        production_db=args.production_db,
-        dashboard_dir=args.dashboard_dir,
-        simulate=args.simulate,
-    )
-    raise SystemExit(0 if success else 1)
